requirements.txt
requests
beautifulsoup4
python-telegram-bot
opencv-python
pandas